refactor longform implementation
