+++ uqlm/long_form/decomposition/__init__.py
@@ -11,3 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from uqlm.long_form.decomposition.response_decomposer import ResponseDecomposer
+
+__all__ = ["ResponseDecomposer"]
